fix: move internal classes to an internal namespace
